refactor: 이미지 최적화를 위해 png 포맷 -> webp 로 변경
