+++ frontend/src/types/image.d.ts
@@ -4,7 +4,6 @@ declare module '*.svg' {
   const ReactComponent: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
   export default ReactComponent;
 }
-
 declare module '*.png';
 declare module '*.webp';
 declare module '*.mp4';
